fix(compiler/abi): LLVM fail because size&align were in bits instead of bytes
