+++ TODO.md
@@ -1,0 +1,6 @@
+# TODO
+
+1. Separate documentation for `TIDEC_LOG` usage with filters
+2. Add tests
+3. Implement a derive macro for deriving spans for a struct
+
