+++ compiler/tidec/src/main.rs
@@ -8,7 +8,7 @@ use tidec_abi::CodegenBackend;
 use tidec_codegen_llvm::builder::CodegenBuilder;
 use tidec_codegen_llvm::context::CodegenCtx;
 use tidec_codegen_llvm::lir::types::BasicTypesUtils;
-use tidec_codegen_llvm::CodegenMethods;
+use tidec_codegen_llvm::traits::CodegenMethods;
 use tidec_lir::lir::LirTyCtx;
 use tidec_lir::syntax::LirTy;
 use tidec_utils::v_debug;
@@ -60,7 +60,6 @@ fn main() {
         .unwrap();
     // module.print_to_stderr();
 
-
     // =========================
     // ========= TESTS =========
     // =========================
@@ -69,7 +68,6 @@ fn main() {
     let align = int_value.get_type().get_alignment();
     println!("Size of i8: {}", int_value);
     println!("Alignment of i8: {}", align);
-
 }
 
 /// Initialize the logger for the tidec project.
