+++ compiler/tidec_abi/Cargo.toml
@@ -5,4 +5,6 @@ edition = "2024"
 
 [dependencies]
 # tidy-alphabetical-start
+tracing = "0.1.41"
 # tidy-alphabetical-end
+
