+++ compiler/tidec_codegen_llvm/Cargo.toml
@@ -10,6 +10,5 @@ tidec_abi = { path = "../tidec_abi" }
 tidec_lir = { path = "../tidec_lir" }
 tidec_utils = { path = "../tidec_utils" }
 tracing = "0.1.41"
-tracing-attributes = "0.1.28"
 # tidy-alphabetical-end
